+++ chain/gssmr/config.toml
@@ -21,7 +21,6 @@ key = ""
 unlock = ""
 
 [core]
-authority = true
 roles = 4
 babe-authority = true
 grandpa-authority = true
